admin-ui/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1
httpx==0.28.1
PyYAML==6.0.2
